# [Perf] set `Symbol.isConcatSpreadable` only when required 