@ljharb 2023-09-05
Is the performance issue caused by *any* assignment of `Symbol.isConcatSpreadable`?

If we `delete empty[Symbol.isConcatSpreadable]`, does that restore the performance?