@infinity-naveen 2023-09-05
> Is the performance issue caused by _any_ assignment of `Symbol.isConcatSpreadable`?
Yes

> If we `delete empty[Symbol.isConcatSpreadable]`, does that restore the performance?

No, that is the issue in V8, if we ever set this symbol once in code, it will impact whole subsequent functionality of concat(). Deleting it will also not restore performance.