@ljharb 2023-09-05
wow, that's horrific. I'll try to poke some v8 folks about it.

this seems viable, then.