@infinity-naveen 2023-09-05
In provided sample script in description, instead of requiring safe-concat-array package, you can just use below line & test:
```
[][Symbol.isConcatSpreadable] = true; // false value will also impact
```

This will also impact performance. Any further mutation in array/value will not restore performance.