@infinity-naveen 2023-09-05
Also, if you try same code in Safari javascript console, it will not impact performance (since, Safari uses different Javascript engine, V8 is used in NodeJs & Chrome)