@ljharb 2023-09-07
It may be expected by v8 devs, but it’s not expected by anyone else :-) a slowdown on operations involving things that set the symbol is expected; a global permanent slowdown on unrelated operations just because it’s assigned anywhere is not.